Move internal/x/wkt to internal/wkt
