+++ internal/x/format/first_pass_visitor.go
@@ -25,9 +25,9 @@ import (
 	"strings"
 
 	"github.com/emicklei/proto"
+	"github.com/uber/prototool/internal/wkt"
 	"github.com/uber/prototool/internal/x/settings"
 	"github.com/uber/prototool/internal/x/text"
-	"github.com/uber/prototool/internal/x/wkt"
 )
 
 type firstPassVisitor struct {
